requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.28.0
Babel==2.13.1
bcrypt==4.0.1
certifi==2023.7.22
click==8.1.7
Django==4.2.6
django-cors-headers==4.3.0
django-decouple==2.1
djangorestframework==3.14.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.10
packaging==23.2
pendulum==2.1.2
Pillow==10.1.0
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
whitenoise==6.6.0
